+++ src/Components/Home/HomeBox.stories.js
@@ -1,8 +1,6 @@
 import React from 'react'
 import {storiesOf} from '@storybook/react'
 
-import {Link} from 'gatsby'
-
 import HomeBox from './HomeBox'
 
 storiesOf('Home/', module).add('HomeBox', () => <HomeBox />)
