+++ src/Components/Layout/Navigation.stories.js
@@ -1,0 +1,6 @@
+import React from 'react'
+import {storiesOf} from '@storybook/react'
+
+import Nav from './Navigation'
+
+storiesOf('Nav/', module).add('Nav', () => <Nav />)
