+++ src/Components/elements/layout/Nav.js
@@ -1,3 +1,5 @@
+import {Link as GLink} from 'gatsby'
+
 import styled, {css} from 'styled-components'
 import {FONT, COOL_GREY_LIGHT} from '../../../config/styles'
 
@@ -6,14 +8,20 @@ const listStyle = css`
 	text-align: right;
 	font-size: 1rem;
 	line-height: 1.4rem;
-	font: ${FONT};
+	font-family: ${FONT};
 	font-weight: 400;
 	color: black;
 `
 
 const Head = styled.div`
 	position: relative;
-	width: 100%;
+	top: 0px;
+	right: 0px;
+	box-sizing: content-box;
+	padding: 0;
+	margin: 0px;
+
+	width: calculate(100% + 200px);
 	height: 100px;
 	margin-bottom: 28px;
 	background-color: ${COOL_GREY_LIGHT};
@@ -21,9 +29,8 @@ const Head = styled.div`
 const Container = styled.nav`
 	position: absolute;
 	z-index: 100;
-	top: 16px;
+	top: 11px;
 	right: 40px;
-	${listStyle};
 `
 const List = styled.ul`
 	${listStyle};
@@ -31,6 +38,20 @@ const List = styled.ul`
 const Item = styled.li`
 	${listStyle};
 	${props => props.home && `font-weight: 600;`}
+	:hover {
+		font-size: 1.2rem;
+		line-height: 1.7rem;
+		font-weight: calculate(font-weight * 2);
+		transition-property: all;
+		transition-duration: 160ms;
+		transition-timing-function: ease-in-out;
+	}
+`
+/* background-color: black; */
+/* color: ${COOL_GREY_LIGHT}; */
+const Link = styled(GLink)`
+	text-decoration: none;
+	box-shadow: none;
 `
 
 const Nav = {}
@@ -38,5 +59,6 @@ Nav.Head = Head
 Nav.Container = Container
 Nav.List = List
 Nav.Item = Item
+Nav.Link = Link
 
 export {Nav}
