+++ src/pages/characters.js
@@ -1,12 +1,35 @@
 import React from 'react'
 import {graphql} from 'gatsby'
-import 'bootstrap/dist/css/bootstrap.css'
+import {Grid, Col, Row} from 'react-styled-flexboxgrid'
 
-import Characters from '../Components/Characters'
+import {CharacterCard} from '../Components/elements/cards'
+import Layout from './../Components/Layout'
 
-const CharactersPage = ({data}) => <Characters data={data} />
+import useRandomCharacters from '../utils/hooks/randomCharacters'
 
-export default CharactersPage
+const Characters = ({data}) => {
+	const characters = useRandomCharacters(data.allCharacters.nodes, 20)
+
+	return (
+		<Layout helmet={'Character Viewer'}>
+			<Grid>
+				<Row>
+					{characters.map((character, index) => (
+						<Col xs={6} sm={4} md={4} lg={3} key={character.id}>
+							<CharacterCard
+								key={character.id}
+								character={character}
+								linkURL={`/characters/${character.id}/`}
+							/>
+						</Col>
+					))}
+				</Row>
+			</Grid>
+		</Layout>
+	)
+}
+
+export default Characters
 
 export const query = graphql`
 	query CharactersQuery {
